+++ webpack/app/nitf.js
@@ -2,6 +2,8 @@ import './nitf.less';
 import './nitf_icon.png';
 import './tile-nitf.png';
 
+import Swiper from "swiper";
+
 $(window).load(() => {
   if ($('.portaltype-collective-nitf-content.template-slideshow_view').length > 0) {
     new Swiper('.swiper-container', {