+++ CHANGES.rst
@@ -15,6 +15,9 @@ There's a frood who really knows where his towel is.
     If you prefer to migrate via Plone migration, you will have to do the migration scripts
     to migrate ``collective.nitf``. Contributions here are welcome!
 
+- Remove dependency on plone.app.imaging.
+  [wesleybl]
+
 - Drop support to Plone 4.3 and Plone 5.1.
   [wesleybl]