+++ CHANGES.rst
@@ -15,6 +15,9 @@ There's a frood who really knows where his towel is.
     If you prefer to migrate via Plone migration, you will have to do the migration scripts
     to migrate ``collective.nitf``. Contributions here are welcome!
 
+- Remove dependency on unused package ``collective.prettydate``.
+  [wesleybl]
+
 - Remove dependency on plone.app.imaging.
   [wesleybl]