+++ CHANGES.rst
@@ -15,7 +15,10 @@ There's a frood who really knows where his towel is.
     If you prefer to migrate via Plone migration, you will have to do the migration scripts
     to migrate ``collective.nitf``. Contributions here are welcome!
 
-- Register static resources in Resource Registry.
+- Load Swiper resources without Resource Registry. Resources are loaded on the Nitf view page.
+  [wesleybl]
+
+- Register internal static resources in Resource Registry.
   [wesleybl]
 
 - Add suport to Python 3.7 and 3.8.